- Fixed the spawn egg fixers not happening - Fixed lotr horses not converting over (not yet fully tested) - Added a cache to PrintLine to make the console a lot less cluttered with debug on, toggleable in Settings - Added a nameFixer function to PlayerData and call this to prevent custom names getting cut off (yet to be tested)
Next up:
- Testing custom item names
- Enchanted items,
- Coloured items,
- Potions

Additional:
In 1.18 the chunk storage is changing slightly again, Old worlds will however be compatible
I'll probably use whichever format is easiest to implement when I get started on the world.
It might actually be beneficial to use the old (1.16, not 1.7.10) format, depending on if Mojang implements populated terrain under current terrain. As using the old format will probably trigger that if implemented correctly.