- Fixed the order issue per Section in SectionMapFixer (moved counting positions to before the writing, instead of after writing to a long) - Fixed accidental reversal of terracotta (old hardened clay) & clay blocks in Conversions
Worlds seem to work fine now, I still need to start using the "Add" tag, but I can't really start testing for that until there is more data in Conversions. Someone I know should be working on BlockIdToName, so hopefully I can continue with that soon.

For now, I'm probably gonna work on BlockMappings in Conversions, on TileEntityFixer and on fixing up EntityFixer